Guess the Number (User Guess)
This is very similar to the last Guess the Number but this time we know the number and the computer is guessing it. Similar principle to the previous project but this time had to apply some more logic to guide the computer to find the correct answer.  Have a go, its really fun